+++ src/Digitas.TesteTecnico/Digitas.Quotes.Domain/ValueObjects/Quote.cs
@@ -2,10 +2,10 @@
 
 public class Quote
 {
-    public int UsdValue { get; set; }
+    public decimal UsdValue { get; set; }
     public decimal Amount { get; set; }
 
-    public Quote(decimal amount, int usdValue)
+    public Quote(decimal amount, decimal usdValue)
     {
         Amount = amount;
         UsdValue = usdValue;
